Remove new resource from File menu.
The File menu new project handles creating new root items for the tree.
New resource item on the other hand was not bound to any particular
place, but would create an element in different places of the tree. If
such feature would continue to exist, it could be argued that new suites
should be able to create from the File menu as well. This would be
unnecessary complex and therefore removing the creation of new
resources from File menu.